+++ Extender/superserial/ui/ScanSettingsTab.java
@@ -1,11 +1,11 @@
 /*
 	ScanSettingsTab.java
 	
-	v0.3 (3/10/2016)
+	v0.4 (7/27/2016)
 	
 	UI Component for the "Scan Settings" configuration tab under the SuperSerial tab. Allows the user to set settings related to the Active Scan checks performed by the 
-	SuperSerial-Active extender (scan all parameters option, number of Node download attempts, time to wait between attempts). Also allows users to add new or edit/delete 
-	existing operating commands used during an Active Scan.
+	SuperSerial-Active extender (scan all parameters option, number of Node download attempts, time to wait between attempts). Also allows users to enable/disable the 
+	different ysoserial payload types used during and Active Scan, as well as add/edit/delete operating commands used during scanning.
 */
 
 package superserial.ui;
@@ -31,6 +31,7 @@ import java.util.Hashtable;
 
 import burp.IBurpExtenderCallbacks;
 import burp.PayloadCommandFactory;
+import burp.PayloadTypeFactory;
 import superserial.settings.SuperSerialSettings;
 
 class ScanSettingsTab extends JPanel {
@@ -38,6 +39,7 @@ class ScanSettingsTab extends JPanel {
 	private JCheckBox scanAllField;
 	private JTextField numAttemptsField;
 	private JTextField waitTimeField;
+	private JCheckBox[] typesFields;
 	private JButton cmdUpButton;
 	private JButton cmdDownButton;
 	private JButton cmdAddButton;
@@ -48,13 +50,15 @@ class ScanSettingsTab extends JPanel {
 	//data fields
 	private SuperSerialSettings settings;
 	private PayloadCommandFactory pcf;
+	private PayloadTypeFactory ptf;
 	private IBurpExtenderCallbacks callbacks;
 
 	ScanSettingsTab(IBurpExtenderCallbacks cb) {
-		super(new GridLayout(4,2));
+		super(new GridLayout(5,2));
 		
 		settings = SuperSerialSettings.getInstance();
 		pcf = PayloadCommandFactory.getInstance();
+		ptf = PayloadTypeFactory.getInstance();
 		callbacks = cb;
 		
 		add(new JLabel("Automatically test all parameters: (WARNING: This will significantly increase scan duration!):",SwingConstants.RIGHT));
@@ -119,44 +123,45 @@ class ScanSettingsTab extends JPanel {
 		});
 		add(waitTimeField);
 		
-		//create commands table
-		Hashtable[] cmdHT = pcf.getCommandsArray();
-		cmdTable = new JTable(new CommandTableModel());
-		dtm = (CommandTableModel) cmdTable.getModel();
-		dtm.addTableModelListener(new TableModelListener() {
-			public void tableChanged(TableModelEvent tme) {
-				switch(tme.getType()) {
-					case TableModelEvent.UPDATE:
-						int firstRow = tme.getFirstRow();
-						int lastRow = tme.getLastRow();
-						//callbacks.printError("UPDATE, first row:"+tme.getFirstRow()+", last row:"+tme.getLastRow()+", source: "+tme.getSource().toString());
-						if(firstRow==lastRow) { //only one row updated (command editted)
-							if(lastRow == dtm.getRowCount()-1) {
-								String cmd = (String) dtm.getValueAt(lastRow,0);
-								String os = (String) dtm.getValueAt(lastRow,1);
-								boolean upload = (Boolean) dtm.getValueAt(lastRow,2);
-								if(lastRow>=pcf.getCommandsCount()) { //add new command
-									pcf.add(cmd,os,"web",upload);
-								} else { //edit existing command
-									pcf.edit(lastRow,cmd,os,"web",upload);
-								}
-							} else {
-								String cmd = (String) dtm.getValueAt(lastRow,0);
-								String os = (String) dtm.getValueAt(lastRow,1);
-								boolean upload = (Boolean) dtm.getValueAt(lastRow,2);
-								pcf.edit(lastRow,cmd,os,"web",upload);
-							}
-						}
-						break;
+		//create payload types checkbox grid
+		int typeCount = ptf.getTypesCount();
+		String[] types = ptf.getAllTypes();
+		int rowCount = (typeCount/4);
+		int lastRowCount = typeCount%4;
+		if(lastRowCount>0) rowCount++;
+		GridLayout tcpGrid = new GridLayout(rowCount,4);
+		JPanel typesCheckPanel = new JPanel(tcpGrid);
+		typesFields = new JCheckBox[types.length];
+		ActionListener typeCheckBoxAL = new ActionListener() {
+			public void actionPerformed(ActionEvent ae) {
+				JCheckBox jcb = (JCheckBox) ae.getSource();
+				ptf.toggleType(jcb.getText(),jcb.isSelected());
+			}
+		};
+		int l=0;
+		//create checkboxes
+		for(int j=0;j<typesFields.length;j++) {
+			typesFields[j] = new JCheckBox(types[l++],false);
+			typesFields[j].addActionListener(typeCheckBoxAL);
+		}
+		//check boxes for enabled types
+		types = ptf.getEnabledTypes();
+		for(int j=0;j<types.length;j++) {
+			for(int k=0;k<typesFields.length;k++) {
+				if(typesFields[k].getText().equals(types[j])) {
+					typesFields[k].setSelected(true);
+					break;
 				}
 			}
-		});
-		
-		//populate table with default built-in commands
-		for(int i=0;i<cmdHT.length;i++) {
-			Hashtable cmd = cmdHT[i];
-			dtm.addRow(new Object[] {cmd.get("cmd"),cmd.get("os"),new Boolean(Boolean.parseBoolean((String) cmd.get("upload")))});
 		}
+		//populate grid; TODO: sort alphetically by columns instead of rows
+		for(int j=0;j<typesFields.length;j++) {
+			typesCheckPanel.add(typesFields[j]);
+		}
+		
+		//add components to panel
+		add(new JLabel("Enabled ysoserial Payload Types:",SwingConstants.RIGHT));
+		add(typesCheckPanel);
 		
 		//create panel for buttons
 		JPanel cmdButtonPanel = new JPanel(new GridLayout(4,1,0,2));
@@ -164,7 +169,7 @@ class ScanSettingsTab extends JPanel {
 		cmdDownButton = new JButton("Move Selected Command Down");
 		cmdAddButton = new JButton("Add New Command");
 		cmdRemoveButton = new JButton("Remove Selected Command(s)");
-		ActionListener buttonAL = new ActionListener() {
+		ActionListener cmdButtonAL = new ActionListener() {
 			public void actionPerformed(ActionEvent ae) {
 				JButton jb = (JButton) ae.getSource();
 				String buttonText = jb.getText();
@@ -173,17 +178,23 @@ class ScanSettingsTab extends JPanel {
 				int rowCount = cmdTable.getRowCount();
 				
 				if(buttonText.contains("Add")) { //add new command (if blank row is not already created)
-					//check if last row command is blank
-					String cmd = (String) dtm.getValueAt(rowCount-1,0);
-					if(cmd!=null) cmd = cmd.trim();
-					int select = rowCount;
-					if((cmd!=null) && (!cmd.isEmpty())) { //if command is inputted
+					//check if table contains rows and if last row command is blank
+					if(dtm.getRowCount()>0) {
+						String cmd = (String) dtm.getValueAt(rowCount-1,0);
+						if(cmd!=null) cmd = cmd.trim();
+						int select = rowCount;
+						if((cmd!=null) && (!cmd.isEmpty())) { //if command is inputted
+							dtm.addRow(new Object[] {"","Unknown",false});
+						} else { //if command is blank
+							select--;
+						}
+						cmdTable.setRowSelectionInterval(select,select);
+						cmdTable.setColumnSelectionInterval(0,0);
+					} else { //if table is blank, add new row
 						dtm.addRow(new Object[] {"","Unknown",false});
-					} else { //if command is blank
-						select--;
+						cmdTable.setRowSelectionInterval(0,0);
+						cmdTable.setColumnSelectionInterval(0,0);
 					}
-					cmdTable.setRowSelectionInterval(select,select);
-					cmdTable.setColumnSelectionInterval(0,0);
 				} else if(selectedRows.length!=0) { //if at least one row is selected
 					if(buttonText.contains("Remove")) { //remove the selected rows; if 1 row is selected, selected the next row
 						for(int i=(selectedRows.length-1);i>-1;i--) {
@@ -202,7 +213,7 @@ class ScanSettingsTab extends JPanel {
 								cmdTable.setRowSelectionInterval(selected,selected);
 							}
 						}
-					} else if(selectedRows.length==1) { //if only 1 row is selected //POSSIBLE TODO: still move rows even if multiple (sequential) rows are selected
+					} else if(selectedRows.length==1) { //if only 1 row is selected
 						if(buttonText.contains("Up")) { //if a row besides the top row or bottom row that contains a blank command is selected, move the selected row up one
 							if(selectedRows[0]!=0) {
 								boolean swap = true;
@@ -234,15 +245,54 @@ class ScanSettingsTab extends JPanel {
 				}
 			}
 		};
-		cmdUpButton.addActionListener(buttonAL);
-		cmdDownButton.addActionListener(buttonAL);
-		cmdAddButton.addActionListener(buttonAL);
-		cmdRemoveButton.addActionListener(buttonAL);
+		cmdUpButton.addActionListener(cmdButtonAL);
+		cmdDownButton.addActionListener(cmdButtonAL);
+		cmdAddButton.addActionListener(cmdButtonAL);
+		cmdRemoveButton.addActionListener(cmdButtonAL);
 		cmdButtonPanel.add(cmdUpButton);
 		cmdButtonPanel.add(cmdDownButton);
 		cmdButtonPanel.add(cmdAddButton);
 		cmdButtonPanel.add(cmdRemoveButton);
 		
+		//create commands table
+		Hashtable[] cmdHT = pcf.getCommandsArray();
+		cmdTable = new JTable(new CommandTableModel());
+		dtm = (CommandTableModel) cmdTable.getModel();
+		dtm.addTableModelListener(new TableModelListener() {
+			public void tableChanged(TableModelEvent tme) {
+				switch(tme.getType()) {
+					case TableModelEvent.UPDATE:
+						int firstRow = tme.getFirstRow();
+						int lastRow = tme.getLastRow();
+						if(firstRow==lastRow) { //only one row updated (command editted)
+							if(lastRow == dtm.getRowCount()-1) {
+								String cmd = (String) dtm.getValueAt(lastRow,0);
+								String os = (String) dtm.getValueAt(lastRow,1);
+								boolean upload = (Boolean) dtm.getValueAt(lastRow,2);
+								if(lastRow>=pcf.getCommandsCount()) { //add new command
+									pcf.add(cmd,os,"web",upload);
+								} else { //edit existing command
+									pcf.edit(lastRow,cmd,os,"web",upload);
+								}
+							} else {
+								String cmd = (String) dtm.getValueAt(lastRow,0);
+								String os = (String) dtm.getValueAt(lastRow,1);
+								boolean upload = (Boolean) dtm.getValueAt(lastRow,2);
+								pcf.edit(lastRow,cmd,os,"web",upload);
+							}
+						}
+						break;
+				}
+			}
+		});
+		
+		//populate table with default built-in commands
+		for(int i=0;i<cmdHT.length;i++) {
+			Hashtable cmd = cmdHT[i];
+			dtm.addRow(new Object[] {cmd.get("cmd"),cmd.get("os"),new Boolean(Boolean.parseBoolean((String) cmd.get("upload")))});
+		}
+		
+		//add buttons and table to panel
 		add(cmdButtonPanel);
 		add(new JScrollPane(cmdTable));
 	}